Refactor: created useFullScreen hook
