+++ src/app/_components/assignments/student/fullscreen-alert.tsx
@@ -15,19 +15,12 @@ interface FullscreenAlertProps {
   onEnterFullscreen: () => void;
 }
 
-export function FullscreenAlert({ onEnterFullscreen }: FullscreenAlertProps) {
+export function FullscreenAlert() {
   const handleEnterFullscreen = () => {
     if (document.documentElement.requestFullscreen) {
-      document.documentElement
-        .requestFullscreen()
-        .then(() => {
-          onEnterFullscreen();
-        })
-        .catch((err) => {
-          console.error(
-            `Error attempting to enable fullscreen: ${err.message}`,
-          );
-        });
+      document.documentElement.requestFullscreen().catch((err) => {
+        console.error(`Error attempting to enable fullscreen: ${err.message}`);
+      });
     }
   };
