+++ src/hooks/use-fullscreen.ts
@@ -1,0 +1,20 @@
+import { useEffect, useState } from "react";
+
+export function useFullScreen() {
+  const [isFullscreen, setIsFullscreen] = useState<boolean>(false);
+  useEffect(() => {
+    const handleFullscreenChange = () => {
+      setIsFullscreen(!!document.fullscreenElement);
+    };
+
+    document.addEventListener("fullscreenchange", handleFullscreenChange);
+
+    setIsFullscreen(!!document.fullscreenElement);
+
+    return () => {
+      document.removeEventListener("fullscreenchange", handleFullscreenChange);
+    };
+  }, []);
+
+  return { isFullscreen };
+}
